🏗️ - Structure as module
