+++ rtrfm2spot/soundcloud_tools.py
@@ -1,0 +1,3 @@
+#NOTE futures
+def search_soundcloud():
+    return NotImplementedError
